refactor(CardItem, Tags, IconCount): 중앙정렬, 문서화
- 모듈명 Card에서 CardItem으로 변경
    - 스타일링 정렬합니다.
    - 사용예시를 문서화합니다.
- Tags를 수정합니다.
    - 가운데 정렬합니다.
    - 많아지면 줄바꿈 처리합니다.
    - tags를 추가 작업을 문서화합니다.
- IconCount 컴포넌트 텍스트를 중앙정렬합니다.

#8
