+++ Components/Common/Card/IconCount.tsx
@@ -19,6 +19,7 @@ const IconCountContainer = styled.div`
   display: flex;
   flex-direction: row;
   align-items: center;
+  font-size: 0.75rem;
   gap: 0;
 `;
