+++ Components/Common/Tags.tsx
@@ -8,6 +8,8 @@ import styled from "styled-components";
  * @TODO size: "sm" | "lg" | "md"; 추가하기
  * 동작하는 로직을 추가할 때는 여기로 추가합니다.
  * 아이콘 클릭에 삭제, 태그 클릭에 추가...
+ *
+ * @TODO 태그 길이 초과하면 줄바꿈 처리 안 하고 (...) 표시하기
  */
 
 interface TagsProps {
@@ -31,6 +33,8 @@ const Tags = ({ tagItems }: TagsProps) => {
 const TagsContainer = styled.ul`
   display: flex;
   gap: 0.25rem;
+  /* 태그를 많이 넣었을 때 줄바꿈 처리합니다. */
+  flex-wrap: wrap;
 `;
 
 type TagsItemType = {
@@ -41,7 +45,7 @@ const TagsItem = styled.li<TagsItemType>`
   background-color: #999999;
   color: #f2f2f2;
   font-size: 0.75rem;
-  padding: 0 0.5rem 0.25rem;
+  padding: 0.25rem 0.5rem;
   border-radius: 0.75rem;
 `;
